Create scrapbox test module
